requirements.txt
Django==1.10
django-admin-views==0.8.0
django-adminplus==0.5
django-ajax-selects==1.5.0
django-audit-log==0.7.0
django-auditlog==0.4.0
django-autocomplete-light==3.1.8
django-cart==1.0.4
django-crispy-forms==1.6.0
django-excel==0.0.4
django-field-history==0.5.0
django-haystack==2.5.0
django-jsonfield==1.0.1
django-nested-inline==0.3.6
django-popup-forms==1.0.3
django-pyodbc-azure==1.10.0.1
django-search==1.0
django-simple-history==1.8.1
elasticsearch==2.4.0
ez-setup==0.9
future==0.15.2
modernize==0.5
Pillow==3.3.1
pyexcel==0.2.4
pyexcel-io==0.2.1
pyexcel-webio==0.0.7
pyexcel-xls==0.2.1
pyodbc==3.0.10
pysolr==3.5.0
pytz==2016.6.1
requests==2.11.0
solr==0.4
texttable==0.8.4
urllib3==1.16
virtualenv==15.0.3
virtualenvwrapper-win==1.2.1
Whoosh==2.7.4
xlrd==1.0.0
xlwt-future==0.8.0